+++ src/utils/core.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@jest/globals';
-import { isCamp, isHQ } from './core';
+import { isCamp, isHQ, isValidRow, isValidCol, isRailroad } from './core';
 
 describe('isCamp', () => {
     test('[2, 1] should be a camp', () => expect(isCamp(2, 1)).toBe(true));
@@ -19,4 +19,70 @@ describe('isHQ', () => {
     test('[0, 3] should be a HQ', () => expect(isHQ(0, 3)).toBe(true));
     test('[11, 1] should be a HQ', () => expect(isHQ(11, 1)).toBe(true));
     test('[11, 3] should be a HQ', () => expect(isHQ(11, 3)).toBe(true));
+});
+
+describe('isValidRow', () => {
+    test('-1 should not be a valid row index', () =>
+        expect(isValidRow(-1)).toBe(false));
+    test('0 should be a valid row index', () =>
+        expect(isValidRow(0)).toBe(true));
+    test('5 should be a valid row index', () =>
+        expect(isValidRow(5)).toBe(true));
+    test('11 should be a valid row index', () =>
+        expect(isValidRow(11)).toBe(true));
+    test('12 should not be a valid row index', () =>
+        expect(isValidRow(12)).toBe(false));
+});
+
+describe('isValidCol', () => {
+    test('-1 should not be a valid column index', () =>
+        expect(isValidCol(-1)).toBe(false));
+    test('0 should be a valid column index', () =>
+        expect(isValidCol(0)).toBe(true));
+    test('4 should be a valid column index', () =>
+        expect(isValidCol(4)).toBe(true));
+    test('5 should be a valid column index', () =>
+        expect(isValidCol(5)).toBe(false));
+});
+
+describe('isRailroad', () => {
+    const generateRow = (r: number) => [...Array(4).keys()].map((c) => [r, c]);
+
+    const topRailCoords = generateRow(1);
+    test('[1, 0] to [1, 4] should be valid railroad coordinates', () =>
+        expect(
+            topRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
+        ).toBe(true));
+
+    const topMidRailCoords = generateRow(5);
+    test('[5, 0] to [5, 4] should be valid railroad coordinates', () =>
+        expect(
+            topMidRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
+        ).toBe(true));
+
+    const botMidRailCoords = generateRow(6);
+    test('[6, 0] to [6, 4] should be valid railroad coordinates', () =>
+        expect(
+            botMidRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
+        ).toBe(true));
+
+    const botRailCoords = generateRow(10);
+    test('[10, 0] to [10, 4] should be valid railroad coordinates', () =>
+        expect(
+            botRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
+        ).toBe(true));
+
+    const generateCol = (c: number) => [...Array(12).keys()].map((r) => [r, c]);
+
+    const leftCol = generateCol(0).slice(1, -1);
+    test('[0, 1] to [0, 11] should be valid railroad coordinates', () =>
+        expect(leftCol.map(([r, c]) => isRailroad(r, c)).every((v) => v)).toBe(
+            true,
+        ));
+
+    const rightCol = generateCol(4).slice(1, -1);
+    test('[4, 1] to [4, 11] should be valid railroad coordinates', () =>
+        expect(rightCol.map(([r, c]) => isRailroad(r, c)).every((v) => v)).toBe(
+            true,
+        ));
 });
