+++ src/utils/getSuccessors.test.ts
@@ -1,90 +1,20 @@
-import { before } from 'node:test';
-import { Board } from './board';
-import { emptyBoard } from './core';
+import { Board, emptyBoard } from './board';
 import {
     getSuccessors,
-    isValidRow,
-    isValidCol,
-    isRailroad,
     isValidDestination,
-    placePiece,
     generateAdjList,
+    _getEngineerRailroadMoves,
 } from './getSuccessors';
 
-import { createPiece } from './piece';
+import { createPiece, placePiece } from './piece';
 
-describe('isValidRow', () => {
-    test('-1 should not be a valid row index', () =>
-        expect(isValidRow(-1)).toBe(false));
-    test('0 should be a valid row index', () =>
-        expect(isValidRow(0)).toBe(true));
-    test('5 should be a valid row index', () =>
-        expect(isValidRow(5)).toBe(true));
-    test('11 should be a valid row index', () =>
-        expect(isValidRow(11)).toBe(true));
-    test('12 should not be a valid row index', () =>
-        expect(isValidRow(12)).toBe(false));
-});
-
-describe('isValidCol', () => {
-    test('-1 should not be a valid column index', () =>
-        expect(isValidCol(-1)).toBe(false));
-    test('0 should be a valid column index', () =>
-        expect(isValidCol(0)).toBe(true));
-    test('4 should be a valid column index', () =>
-        expect(isValidCol(4)).toBe(true));
-    test('5 should be a valid column index', () =>
-        expect(isValidCol(5)).toBe(false));
-});
-
-describe('isRailroad', () => {
-    const generateRow = (r: number) => [...Array(4).keys()].map((c) => [r, c]);
-
-    const topRailCoords = generateRow(1);
-    test('[1, 0] to [1, 4] should be valid railroad coordinates', () =>
-        expect(
-            topRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
-        ).toBe(true));
-
-    const topMidRailCoords = generateRow(5);
-    test('[5, 0] to [5, 4] should be valid railroad coordinates', () =>
-        expect(
-            topMidRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
-        ).toBe(true));
-
-    const botMidRailCoords = generateRow(6);
-    test('[6, 0] to [6, 4] should be valid railroad coordinates', () =>
-        expect(
-            botMidRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
-        ).toBe(true));
-
-    const botRailCoords = generateRow(10);
-    test('[10, 0] to [10, 4] should be valid railroad coordinates', () =>
-        expect(
-            botRailCoords.map(([r, c]) => isRailroad(r, c)).every((v) => v),
-        ).toBe(true));
-
-    const generateCol = (c: number) => [...Array(12).keys()].map((r) => [r, c]);
-
-    const leftCol = generateCol(0).slice(1, -1);
-    test('[0, 1] to [0, 11] should be valid railroad coordinates', () =>
-        expect(leftCol.map(([r, c]) => isRailroad(r, c)).every((v) => v)).toBe(
-            true,
-        ));
-
-    const rightCol = generateCol(4).slice(1, -1);
-    test('[4, 1] to [4, 11] should be valid railroad coordinates', () =>
-        expect(rightCol.map(([r, c]) => isRailroad(r, c)).every((v) => v)).toBe(
-            true,
-        ));
-});
 
 describe('isValidDestination', () => {
     let board = emptyBoard();
 
     beforeEach(() => {
         board = emptyBoard();
-    })
+    });
 
     describe('test out of bound destinations', () => {
         test('out of bound x move [-1, 5] should not be valid', () => {
@@ -98,7 +28,7 @@ describe('isValidDestination', () => {
     describe('test normal position occupation by friendly vs enemy', () => {
         beforeEach(() => {
             board = placePiece(board, 0, 0, createPiece('field_marshall', 0));
-        })
+        });
         test('[0, 0] should be an invalid destination for affiliation 0 since you cannot attack your own troops', () => {
             expect(isValidDestination(board, 0, 0, 0)).toBe(false);
         });
@@ -123,6 +53,81 @@ describe('isValidDestination', () => {
         });
     });
 });
+
+describe('_getEngineerRailroadMoves', () => {
+    let board = emptyBoard();
+
+    beforeEach(() => {
+        board = emptyBoard();
+    });
+
+    test('Engineer should be able to visit all railroad tracks as long as there are no obstructions by other pieces', () => {
+        board = placePiece(board, 1, 0, createPiece('engineer', 0));
+        const moves = _getEngineerRailroadMoves(board, 1, 0, 0);
+        // engineer's own position should not be included
+        expect(moves.has(JSON.stringify([1, 0]))).toBe(false);
+
+        // left vertical track (excluding engineer's current position)
+        for (let i = 2; i < 11; i++) {
+            expect(moves.has(JSON.stringify([i, 0]))).toBe(true);
+        }
+
+        // right vertical track (all)
+        for (let i = 1; i < 11; i++) {
+            expect(moves.has(JSON.stringify([i, 4]))).toBe(true);
+        }
+
+        // horizontal track at row 1 (excluding engineer's current position)
+        for (let i = 1; i < 5; i++) {
+            expect(moves.has(JSON.stringify([1, i]))).toBe(true);
+        }
+
+        // horizontal track at row 5 (all)
+        for (let i = 0; i < 5; i++) {
+            expect(moves.has(JSON.stringify([5, i]))).toBe(true);
+        }
+
+        // horizontal track at row 6 (all)
+        for (let i = 0; i < 5; i++) {
+            expect(moves.has(JSON.stringify([6, i]))).toBe(true);
+        }
+
+        // horizontal track at row 10 (all)
+        for (let i = 0; i < 5; i++) {
+            expect(moves.has(JSON.stringify([10, i]))).toBe(true);
+        }
+    });
+
+    test('Engineer should be blocked by an enemy piece', () => {
+        board = placePiece(board, 1, 0, createPiece('engineer', 0));
+        board = placePiece(board, 1, 2, createPiece('bomb', 1));
+        board = placePiece(board, 1, 4, createPiece('bomb', 1));
+        const moves = _getEngineerRailroadMoves(board, 1, 0, 0);
+
+        expect(moves.has(JSON.stringify([1, 0]))).toBe(false);
+        expect(moves.has(JSON.stringify([1, 1]))).toBe(true);
+        expect(moves.has(JSON.stringify([1, 2]))).toBe(true);
+        expect(moves.has(JSON.stringify([1, 3]))).toBe(false);
+        expect(moves.has(JSON.stringify([1, 4]))).toBe(true);
+        expect(moves.has(JSON.stringify([2, 4]))).toBe(true);
+    });
+
+    test('Engineer should be blocked by an allied piece', () => {
+        board = placePiece(board, 1, 0, createPiece('engineer', 0));
+        board = placePiece(board, 1, 2, createPiece('bomb', 0));
+        board = placePiece(board, 1, 4, createPiece('bomb', 0));
+        const moves = _getEngineerRailroadMoves(board, 1, 0, 0);
+
+        expect(moves.has(JSON.stringify([1, 0]))).toBe(false);
+        expect(moves.has(JSON.stringify([1, 1]))).toBe(true);
+        expect(moves.has(JSON.stringify([1, 2]))).toBe(false);
+        expect(moves.has(JSON.stringify([1, 3]))).toBe(false);
+        expect(moves.has(JSON.stringify([1, 4]))).toBe(false);
+        expect(moves.has(JSON.stringify([2, 4]))).toBe(true);
+    });
+});
+
+
 
 describe('generateAdjList', () => {
     const adjList = generateAdjList();
